requirements.txt
alembic==1.18.4
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
google-auth==2.48.0
google-auth-oauthlib==1.2.4
greenlet==3.3.2
gspread==6.2.1
h11==0.16.0
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
login==0.0.6
Mako==1.3.10
MarkupSafe==3.0.3
msgspec==0.20.0
numpy==2.4.3
oauth2client==4.1.3
oauthlib==3.3.1
packaging==26.0
pandas==3.0.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-engineio==4.13.1
python-socketio==5.16.1
python-dotenv==1.0.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
wsproto==1.3.2
